Client Server Changes
When Driver logs in to app, notifies the server.
Server makes list of connected Drivers.
The server identifies driver and passenger clients.
Server sends back a list of connected valid Drivers to the clients.
It handles properly if a client gets disconnected.
If a driver client gets disconnected, don’t show in the list of drivers.
Passenger sends message to server when making a ride.
Server sends message to Driver.
It returns a list of valid drivers only to passengers.